+++ src/main/java/com/site21/bittermelon/common/systems/atmosphere/AtmosHandler.java
@@ -16,7 +16,10 @@ import net.neoforged.neoforge.network.PacketDistributor;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.*;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+import java.util.UUID;
 
 import static com.site21.bittermelon.init.neoforge.BitterAttachmentTypes.ATMOSPHERE;
 
@@ -166,7 +169,7 @@ public final class AtmosHandler {
             );
 
             // Reduce gases in the original instance proportionally
-            float remainingRatio = (float) newBlocks.size() / atmosInstance.getBlocks().size();
+            int remainingRatio = newBlocks.size() / atmosInstance.getBlocks().size();
             for (SubstanceStack gas : atmosInstance.getGases()) {
                 gas.setAmount(gas.getAmount() * remainingRatio);
             }
